requirements.txt
annotated-types==0.6.0
azure-core==1.30.1
azure-functions==1.18.0
azure-identity==1.15.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.8.0
azure-keyvault-keys==4.9.0
azure-keyvault-secrets==4.8.0
azure-storage-blob==12.19.1
azure-storage-file-datalake==12.14.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coverage==7.4.3
cramjam==2.8.3
cryptography==42.0.5
Faker==24.0.0
fastparquet==2024.2.0
fsspec==2024.3.1
idna==3.6
isodate==0.6.1
msal==1.27.0
msal-extensions==1.1.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
portalocker==2.8.2
pyarrow==14.0.2
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
tradingeconomics==4.3.5
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
websocket-client==1.7.0
